requirements.txt
fastapi
uvicorn[standard]
websockets
python-dotenv
deepgram-sdk
fuzzywuzzy
python-Levenshtein